refactor: collection view 관련 보조 View 및 Cell 파일 분리
