+++ Projects/Features/HomeFeature/Sources/Presentation/HomeSectionHeaderView.swift
@@ -1,0 +1,35 @@
+//
+//  HomeSectionHeaderView.swift
+//  HomeFeature
+//
+//  Created by 박서연 on 1/20/26.
+//  Copyright © 2026 linda. All rights reserved.
+//
+
+import UIKit
+
+final class HomeSectionHeaderView: UICollectionReusableView {
+    private let titleLabel = UILabel()
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        titleLabel.textColor = .white
+        titleLabel.font = .boldSystemFont(ofSize: 18)
+        
+        addSubview(titleLabel)
+        titleLabel.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            titleLabel.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 16),
+            titleLabel.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -16),
+            titleLabel.topAnchor.constraint(equalTo: topAnchor),
+            titleLabel.bottomAnchor.constraint(equalTo: bottomAnchor)
+        ])
+    }
+    
+    required init?(coder: NSCoder) { fatalError() }
+    
+    func configure(title: String) {
+        titleLabel.text = title
+    }
+}
