+++ Projects/Features/HomeFeature/Sources/Presentation/PosterCell.swift
@@ -1,0 +1,60 @@
+//
+//  PosterCell.swift
+//  HomeFeature
+//
+//  Created by 박서연 on 1/20/26.
+//  Copyright © 2026 linda. All rights reserved.
+//
+
+import UIKit
+
+final class PosterCell: UICollectionViewCell {
+    private let imageView = UIImageView()
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        contentView.layer.cornerRadius = 8
+        contentView.layer.masksToBounds = true
+        
+        imageView.contentMode = .scaleAspectFill
+        contentView.addSubview(imageView)
+        imageView.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            imageView.topAnchor.constraint(equalTo: contentView.topAnchor),
+            imageView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+            imageView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+            imageView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor)
+        ])
+    }
+    
+    required init?(coder: NSCoder) { fatalError() }
+    
+    override func prepareForReuse() {
+        super.prepareForReuse()
+        imageView.image = nil
+    }
+    
+    func configure(posterURL: URL?) {
+        guard let url = posterURL else {
+            imageView.backgroundColor = .darkGray
+            return
+        }
+        
+        Task {
+            do {
+                let (data, _) = try await URLSession.shared.data(from: url)
+                if let image = UIImage(data: data) {
+                    await MainActor.run {
+                        self.imageView.image = image
+                    }
+                }
+            } catch {
+                await MainActor.run {
+                    self.imageView.backgroundColor = .darkGray
+                }
+            }
+        }
+    }
+}
+
